requirements.txt
pdf2image
PyMuPDF
transformers
torch
numpy
Pillow
opencv-python
